Clean up Entity

+++ src/main/java/social/alone/server/link/LinkRepository.java
@@ -2,5 +2,5 @@ package social.alone.server.link;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LinkRepository extends JpaRepository<Link, Integer> {
+public interface LinkRepository extends JpaRepository<Link, Long> {
 }
